refactor: use lib not src in test
