+++ package.json
@@ -12,7 +12,7 @@
     "prepare": "npm run compile",
     "mocha": "cross-env TS_NODE_FILES=true mocha",
     "lint": "eslint --ext .ts .",
-    "spec": "npm run mocha -- test/*.ts",
+    "spec": "tsc && npm run mocha -- test/*.ts",
     "test": "npm run lint && npm run spec"
   },
   "bin": {
