+++ test/module-type-node-gyp.ts
@@ -4,8 +4,8 @@ import os from 'os';
 import path from 'path';
 
 import { cleanupTestModule, resetTestModule } from './helpers/module-setup';
-import { NodeGyp } from '../src/module-type/node-gyp/node-gyp';
-import { Rebuilder } from '../src/rebuild';
+import { NodeGyp } from '../lib/module-type/node-gyp/node-gyp';
+import { Rebuilder } from '../lib/rebuild';
 
 describe('node-gyp', () => {
   describe('buildArgs', () => {
