+++ test/read-package-json.ts
@@ -1,7 +1,7 @@
 import * as path from 'path';
 import { expect } from 'chai';
 
-import { readPackageJson } from '../src/read-package-json';
+import { readPackageJson } from '../lib/read-package-json';
 
 describe('read-package-json', () => {
   it('should find a package.json file from the given directory', async () => {
